+++ websocket/example/main.go
@@ -1,0 +1,54 @@
+package main
+
+import (
+	"os"
+	"os/signal"
+
+	polygonws "github.com/polygon-io/client-go/websocket"
+	"github.com/polygon-io/client-go/websocket/models"
+	"github.com/sirupsen/logrus"
+)
+
+func main() {
+	log := logrus.New()
+	log.SetLevel(logrus.DebugLevel)
+	log.SetFormatter(&logrus.JSONFormatter{})
+	c, err := polygonws.New(polygonws.Config{
+		APIKey: os.Getenv("POLYGON_API_KEY"),
+		Feed:   polygonws.RealTime,
+		Market: polygonws.Stocks,
+		Log:    log,
+	})
+	if err != nil {
+		log.Fatal(err)
+	}
+	defer c.Close()
+
+	_ = c.Subscribe(polygonws.StocksTrades, "SPY")
+	_ = c.Subscribe(polygonws.StocksQuotes, "SPY")
+	if err := c.Connect(); err != nil {
+		log.Fatal(err)
+	}
+
+	sigint := make(chan os.Signal, 1)
+	signal.Notify(sigint, os.Interrupt)
+
+	for {
+		select {
+		case <-sigint:
+			return
+		case <-c.Error():
+			return
+		case out, more := <-c.Output():
+			if !more {
+				return
+			}
+			switch out.(type) {
+			case models.EquityTrade:
+				log.WithFields(logrus.Fields{"trade": out}).Info()
+			case models.EquityQuote:
+				log.WithFields(logrus.Fields{"quote": out}).Info()
+			}
+		}
+	}
+}
